nav in graphs bar extracted into seperate component
Nav was extracted for readabilty purposes. Attempting to
separate the concerns in the graph container and increase
understandability
